+++ lib/axn/webhooks.rb
@@ -9,6 +9,7 @@ require_relative "webhooks/signature"
 require_relative "webhooks/resolvers"
 require_relative "webhooks/verify"
 require_relative "webhooks/verifiers"
+require_relative "webhooks/verifiers/hmac"
 require_relative "webhooks/inbound"
 
 module Axn